# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Argus — AI Onchain Intelligence & Whale Tracking Agent API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /wallets:
    get:
      operationId: listWallets
      tags: [wallets]
      summary: List all tracked wallets
      parameters:
        - name: watchlistId
          in: query
          schema:
            type: integer
        - name: category
          in: query
          schema:
            type: string
        - name: chain
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of wallets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Wallet"
    post:
      operationId: createWallet
      tags: [wallets]
      summary: Add a wallet to track
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WalletInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Wallet"
  /wallets/{id}:
    get:
      operationId: getWallet
      tags: [wallets]
      summary: Get wallet by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Wallet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Wallet"
        "404":
          description: Not found
    patch:
      operationId: updateWallet
      tags: [wallets]
      summary: Update wallet metadata
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WalletUpdate"
      responses:
        "200":
          description: Updated wallet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Wallet"
    delete:
      operationId: deleteWallet
      tags: [wallets]
      summary: Remove a tracked wallet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Watchlists ────────────────────────────────────────────────────────────
  /watchlists:
    get:
      operationId: listWatchlists
      tags: [watchlists]
      summary: List all watchlists
      responses:
        "200":
          description: List of watchlists
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Watchlist"
    post:
      operationId: createWatchlist
      tags: [watchlists]
      summary: Create a new watchlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WatchlistInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Watchlist"
  /watchlists/{id}:
    get:
      operationId: getWatchlist
      tags: [watchlists]
      summary: Get watchlist with its wallets
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Watchlist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WatchlistDetail"
        "404":
          description: Not found
    patch:
      operationId: updateWatchlist
      tags: [watchlists]
      summary: Update watchlist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WatchlistUpdate"
      responses:
        "200":
          description: Updated watchlist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Watchlist"
    delete:
      operationId: deleteWatchlist
      tags: [watchlists]
      summary: Delete a watchlist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /watchlists/{id}/wallets:
    post:
      operationId: addWalletToWatchlist
      tags: [watchlists]
      summary: Add wallet to watchlist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WatchlistWalletInput"
      responses:
        "201":
          description: Added
  /watchlists/{id}/wallets/{walletId}:
    delete:
      operationId: removeWalletFromWatchlist
      tags: [watchlists]
      summary: Remove wallet from watchlist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: walletId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed

  # ── Events ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Wallet:
      type: object
      required: [id, address, chain, category, label, isActive, createdAt]
      properties:
        id:         { type: integer }
        address:    { type: string }
        chain:      { type: string }
        category:   { type: string }
        label:      { type: string }
        notes:      { type: ["string", "null"] }
        isActive:   { type: boolean }
        lastActivity: { type: ["string", "null"] }
        totalVolume30d: { type: ["number", "null"] }
        eventCount: { type: integer }
        createdAt:  { type: string }
    WalletInput:
      type: object
      required: [address, chain, category, label]
      properties:
        address:  { type: string, minLength: 10 }
        chain:    { type: string }
        category: { type: string }
        label:    { type: string }
        notes:    { type: string }
    WalletUpdate:
      type: object
      properties:
        label:    { type: string }
        notes:    { type: string }
        category: { type: string }
        isActive: { type: boolean }

    # ── Watchlist ───────────────────────────────────────────────────────────
    Watchlist:
      type: object
      required: [id, name, walletCount, createdAt]
      properties:
        id:          { type: integer }
        name:        { type: string }
        description: { type: ["string", "null"] }
        walletCount: { type: integer }
        createdAt:   { type: string }
    WatchlistDetail:
      type: object
      required: [id, name, wallets, createdAt]
      properties:
        id:          { type: integer }
        name:        { type: string }
        description: { type: ["string", "null"] }
        wallets:
          type: array
          items:
            $ref: "#/components/schemas/Wallet"
        createdAt:   { type: string }
    WatchlistInput:
      type: object
      required: [name]
      properties:
        name:        { type: string, minLength: 1 }
        description: { type: string }
    WatchlistUpdate:
      type: object
      properties:
        name:        { type: string }
        description: { type: string }
    WatchlistWalletInput:
      type: object
      required: [walletId]
      properties:
        walletId: { type: integer }

    # ── Events ──────────────────────────────────────────────────────────────
...