# lib/api-zod/src/generated/types/walletInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Argus — AI Onchain Intelligence & Whale Tracking Agent API
 * OpenAPI spec version: 0.1.0
 */

export interface WalletInput {
  /** @minLength 10 */
  address: string;
  chain: string;
  category: string;
  label: string;
  notes?: string;
}