# lib/api-zod/src/generated/types/watchlistDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Argus — AI Onchain Intelligence & Whale Tracking Agent API
 * OpenAPI spec version: 0.1.0
 */
import type { Wallet } from './wallet';

export interface WatchlistDetail {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  wallets: Wallet[];
  createdAt: string;
}