# lib/api-zod/src/generated/types/watchlistInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Argus — AI Onchain Intelligence & Whale Tracking Agent API
 * OpenAPI spec version: 0.1.0
 */

export interface WatchlistInput {
  /** @minLength 1 */
  name: string;
  description?: string;
}